Only create ssh key if it doesn't exist

+++ app/models/deployment/credentials.rb
@@ -44,8 +44,10 @@ class Deployment
       FileUtils.mkdir_p ssh_directory
       FileUtils.chmod_R 0700, ssh_directory
 
-      File.open(ssh_key, "w", 0600) do |fp|
-        fp.puts(ssh_private_key.split('\n'))
+      if !File.exist?(ssh_key)
+        File.open(ssh_key, "w", 0600) do |fp|
+          fp.puts(ssh_private_key.split('\n'))
+        end
       end
 
       File.open(ssh_config, "w", 0600) do |fp|
